+++ NEWS.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - **README.md**: Use vanilla (not GitHub-flavored) Markdown syntax for snippets
+- Replace (redraw-display) with (force-mode-line-update t)
 
 ### Removed